# packages/api-contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DSHWAR API",
    "version": "0.9.0",
    "description": "DeepSeek Harness 之上的 ToB 产品基座的 API 平面。\n\n**这份契约的稳定性承诺不依赖上游 dsh 版本。** 事件词表、错误码、资源形状\n全部由 DSHWAR 定义并版本化 —— 上游改接口时,受影响的是网关内部的翻译层,\n不是这份契约。\n\n标注 `x-dshwar-status: planned` 的端点契约完整但尚未实现,调用返回 501,\n响应头 `x-dshwar-planned-version` 指出计划版本。它们返回 501 而非 404,\n是为了让第三方知道「这个端点是真的,只是还没到」。",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "{scheme}://{host}",
      "description": "部署方自有域名。TLS 由反向代理终结,网关不自己管证书",
      "variables": {
        "scheme": {
          "default": "https",
          "enum": [
            "https",
            "http"
          ]
        },
        "host": {
          "default": "dshwar.localhost:3000"
        }
      }
    }
  ],
...
  "security": [
    {
      "runtimeBearer": []
    }
  ],
  "paths": {
...
    "/v1/admin/subjects/{id}/credentials": {
      "get": {
        "operationId": "listSubjectCredentials",
        "summary": "凭据配置状态",
        "description": "**永不返回值。** 只暴露 configured / source / writable —— 本端点的 schema 里没有任何可以放值的字段(CLAUDE.md 硬规则 5)。",
        "tags": [
          "admin",
          "credentials"
        ],
        "security": [
          {
            "adminApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "主体标识。必须是 IdP 的不可变主键(OIDC sub / SCIM id / 目录 object id),不得使用邮箱",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "本页最多返回多少条",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "上一页返回的 nextCursor;首页省略",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "排序字段,前缀 `-` 表示降序(如 `-createdAt`)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "凭据配置状态列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListCredentialsResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求体或参数不满足 schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "凭证缺失、无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin Key 不属于该主体所在租户",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在,或存在但不属于该主体(两者刻意不可区分)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "触发限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-dshwar-status": "implemented"
      }
    },
    "/v1/admin/subjects": {
      "get": {
        "operationId": "listSubjects",
        "summary": "列出用户镜像",
        "tags": [
          "admin",
          "subjects"
        ],
        "security": [
          {
            "adminApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "本页最多返回多少条",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "上一页返回的 nextCursor;首页省略",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "排序字段,前缀 `-` 表示降序(如 `-createdAt`)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户镜像列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSubjectsResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求体或参数不满足 schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "凭证缺失、无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "触发限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-dshwar-status": "implemented"
      }
    },
    "/v1/admin/subjects/{id}": {
      "get": {
        "operationId": "getSubject",
        "summary": "读取一个用户镜像",
        "tags": [
          "admin",
          "subjects"
        ],
        "security": [
          {
            "adminApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "主体标识。必须是 IdP 的不可变主键(OIDC sub / SCIM id / 目录 object id),不得使用邮箱",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用户镜像",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSubjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求体或参数不满足 schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "凭证缺失、无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在,或存在但不属于该主体(两者刻意不可区分)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "触发限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-dshwar-status": "implemented"
      }
    },
...
    "/v1/admin/subjects/{id}/usage": {
      "get": {
        "operationId": "listSubjectUsage",
        "summary": "单个主体的用量明细",
        "tags": [
          "admin",
          "usage"
        ],
        "security": [
          {
            "adminApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "主体标识。必须是 IdP 的不可变主键(OIDC sub / SCIM id / 目录 object id),不得使用邮箱",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 256
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "本页最多返回多少条",
            "schema": {
              "default": 50,
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "description": "上一页返回的 nextCursor;首页省略",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "description": "排序字段,前缀 `-` 表示降序(如 `-createdAt`)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用量明细",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListUsageResponse"
                }
              }
            }
          },
          "400": {
            "description": "请求体或参数不满足 schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "凭证缺失、无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在,或存在但不属于该主体(两者刻意不可区分)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "触发限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-dshwar-status": "implemented"
      }
    },
...
    "/v1/admin/capacity": {
      "get": {
        "operationId": "getCapacity",
        "summary": "部署容量:隔离档、进程上限、成员上限",
        "tags": [
          "admin"
        ],
        "security": [
          {
            "adminApiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "当前部署的容量",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capacity"
                }
              }
            }
          },
          "400": {
            "description": "请求体或参数不满足 schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "凭证缺失、无效或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "触发限流",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务端内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-dshwar-status": "implemented"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "runtimeBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "终端用户令牌。由部署方的 IdP 签发,DSHWAR 只验证与映射"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-DSHWAR-Admin-Key",
        "description": "**按租户签发**,一把钥匙不得横跨租户。与运行时令牌分离 —— 供给系统只能写身份镜像,不能读用量与凭据配置"
      },
      "webhookSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "Stripe-Signature",
        "description": "webhook 专用(V0.6.0):凭证是请求体的 HMAC-SHA256 签名,时间戳参与签名以抗重放。无签名或签名无效统一 401,不区分原因"
      }
    },
    "schemas": {
...
      "ListCredentialsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CredentialDescriptor"
            }
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "requestId": {
            "type": "string",
            "description": "与服务端日志和审计记录对得上的调用标识"
          }
        },
        "required": [
          "data",
          "nextCursor",
          "requestId"
        ],
        "additionalProperties": false
      },
...
    }
  }
}